Update embed.min.js
A V1 Riddle that was migrated to V2 but embedded with a V1 embed code caused an infinite loop.

The code was optimized to ensure that an infinite loop can no longer occur under any circumstances.

In the V1 embed code, we now check if a V2 event is being sent. If so, it indicates a migrated Riddle, and it will be initialized as a V2 Riddle.